+++ pmd/src/net/sourceforge/pmd/cpd/CPDNullListener.java
@@ -4,5 +4,5 @@ import java.io.File;
 
 public class CPDNullListener implements CPDListener {
     public void addedFile(int fileCount, File file) {}
-    public void comparisonCountUpdate(int comparisons) {}
+    public void comparisonCountUpdate(long comparisons) {}
 }
